requirements.txt
appdirs==1.4.3
astroid==1.5.2
cffi==1.9.1
click==6.7
conda==4.3.14
cryptography==1.7.1
Flask==0.12.2
idna==2.2
isort==4.2.5
itsdangerous==0.24
jep==3.6.3
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.12.1
packaging==16.8
protobuf==3.3.0
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.17
pylint==1.7.1
pyOpenSSL==16.2.0
pyparsing==2.2.0
PyYAML==3.12
redis==2.10.5
requests==2.12.4
schedule==0.4.2
six==1.10.0
tensorflow==1.1.0
Werkzeug==0.12.2
wrapt==1.10.10
